refactor(messenger): ChatSettings -> ContactSettings
