+++ js/packages/components/chat/OneToOneSettings.tsx
@@ -1,131 +1,131 @@
-import React, { useState } from 'react'
-import { View, ScrollView, ActivityIndicator } from 'react-native'
-import { Text } from 'react-native-ui-kitten'
+import { messenger as messengerpb } from '@berty-tech/api/index.js'
+import { ScreenProps, useNavigation } from '@berty-tech/navigation'
+import { useContact, useConversation } from '@berty-tech/store/hooks'
 import { useStyles } from '@berty-tech/styles'
-import { ButtonSetting } from '../shared-components/SettingsButtons'
-import { FingerprintContent } from '../shared-components/FingerprintContent'
-import { TabBar } from '../shared-components/TabBar'
+import React from 'react'
+import { ScrollView, View } from 'react-native'
+import { Text } from 'react-native-ui-kitten'
 import HeaderSettings from '../shared-components/Header'
-import { useNavigation, ScreenProps } from '@berty-tech/navigation'
 import { ProceduralCircleAvatar } from '../shared-components/ProceduralCircleAvatar'
-import { useContacts } from '@berty-tech/store/hooks'
+import { ButtonSetting, ButtonSettingRow } from '../shared-components/SettingsButtons'
 
-const OneToOneSettingsHeaderContent: React.FC = ({ children }) => {
-	const [{ margin }] = useStyles()
-	return <View style={[margin.top.big]}>{children}</View>
-}
+//
+// OneToOneSettings
+//
 
-const SelectedContent: React.FC<{ contentName: string; publicKey: string }> = ({
-	contentName,
-	publicKey,
-}) => {
-	switch (contentName) {
-		case 'Fingerprint':
-			return <FingerprintContent seed={publicKey} />
-		default:
-			return <Text>Error: Unknown content name "{contentName}"</Text>
+// Styles
+const useStylesOneToOne = () => {
+	const [{ flex, height, margin }] = useStyles()
+	return {
+		headerAvatar: flex.large,
+		firstHeaderButton: [margin.right.scale(20), height(90)],
+		secondHeaderButton: [margin.right.scale(20), height(90)],
+		thirdHeaderButton: height(90),
 	}
 }
 
-const OneToOneSettingsHeader: React.FC<{ contact: any }> = ({ contact }) => {
-	const [{ border, background, padding, row, absolute, text }] = useStyles()
-	const [selectedContent, setSelectedContent] = useState('Fingerprint')
+const OneToOneHeader: React.FC<{ contact: any }> = ({ contact }) => {
+	const _styles = useStylesOneToOne()
+	const [{ text, padding, border, row }] = useStyles()
 	return (
-		<View style={[padding.medium, padding.top.scale(50)]}>
-			<View
-				style={[
-					border.radius.scale(30),
-					background.white,
-					padding.horizontal.medium,
-					padding.bottom.medium,
-				]}
+		<View style={[_styles.headerAvatar]}>
+			<ProceduralCircleAvatar
+				seed={contact.publicKey}
+				style={[border.shadow.big, row.center]}
+				diffSize={30}
+			/>
+			<Text
+				numberOfLines={1}
+				style={[text.size.scale(18), text.color.white, text.align.center, padding.top.small]}
 			>
-				<View style={[row.item.justify, absolute.scale({ top: -50 })]}>
-					<ProceduralCircleAvatar
-						seed={contact.publicKey}
-						style={[border.shadow.big]}
-						diffSize={30}
-					/>
-				</View>
-				<View style={[padding.horizontal.medium, padding.bottom.medium, padding.top.scale(65)]}>
-					<Text style={[text.size.big, text.color.black, text.align.center, text.bold.small]}>
-						{contact.displayName}
-					</Text>
-					<TabBar
-						tabs={[
-							{ name: 'Fingerprint', icon: 'fingerprint', iconPack: 'custom' },
-							{ name: 'Infos', icon: 'info-outline', buttonDisabled: true },
-							{
-								name: 'Devices',
-								icon: 'smartphone',
-								iconPack: 'feather',
-								iconTransform: [{ rotate: '22.5deg' }, { scale: 0.8 }],
-								buttonDisabled: true,
-							},
-						]}
-						onTabChange={setSelectedContent}
-					/>
-					<OneToOneSettingsHeaderContent>
-						<SelectedContent publicKey={contact.publicKey} contentName={selectedContent} />
-					</OneToOneSettingsHeaderContent>
-				</View>
-			</View>
+				{contact.displayName}
+			</Text>
 		</View>
 	)
 }
 
-const DeleteContactButton: React.FC<{ id: string }> = ({ id }) => {
-	const [{ color }] = useStyles()
-	// const deleteContact = Messenger.useDeleteContact()
-	const deleteContact = () => {}
-	return (
-		<ButtonSetting
-			name='Delete contact'
-			icon='trash-2-outline'
-			iconColor={color.red}
-			onPress={() => deleteContact({ id })}
-		/>
-	)
-}
-
-const OneToOneSettingsBody: React.FC<{ id: string }> = ({ id }) => {
+const OneToOneHeaderButtons: React.FC<{}> = () => {
+	const _styles = useStylesOneToOne()
 	const [{ padding, color }] = useStyles()
 	return (
-		<View style={padding.medium}>
-			<ButtonSetting
-				icon='checkmark-circle-2'
-				name='Mark as verified'
-				iconDependToggle
-				toggled
-				disabled
+		<View style={[padding.horizontal.medium, padding.top.medium]}>
+			<ButtonSettingRow
+				state={[
+					{
+						name: 'Search',
+						icon: 'search-outline',
+						color: color.blue,
+						style: _styles.firstHeaderButton,
+						disabled: true,
+					},
+					{
+						name: 'Call',
+						icon: 'phone-outline',
+						color: color.green,
+						style: _styles.secondHeaderButton,
+						disabled: true,
+					},
+					{
+						name: 'Share',
+						icon: 'upload',
+						color: color.blue,
+						style: _styles.thirdHeaderButton,
+						disabled: true,
+					},
+				]}
 			/>
-			<ButtonSetting name='Block contact' icon='slash-outline' iconColor={color.red} disabled />
-			<DeleteContactButton id={id} />
 		</View>
 	)
 }
 
-export const OneToOneSettings: React.FC<ScreenProps.Chat.OneToOneSettings> = ({ route }) => {
-	const { contactId } = route.params
-	const { goBack } = useNavigation()
-	const contact: any = (useContacts() as any)[contactId] || null
-	const [{ background, flex }] = useStyles()
-	if (!contact) {
+const OneToOneBody: React.FC<{}> = () => {
+	const [{ padding, color }] = useStyles()
+	return (
+		<View style={[padding.horizontal.medium]}>
+			<ButtonSetting name='Medias, links & docs' icon='image-outline' disabled />
+			<ButtonSetting name='Receive notifications' icon='bell-outline' toggled disabled />
+			<ButtonSetting
+				name='Mutual groups'
+				icon='users'
+				iconPack='custom'
+				state={{ value: '3 mutuals', color: color.blue, bgColor: color.light.blue }}
+				disabled
+			/>
+			<ButtonSetting
+				name='Erase conversation'
+				icon='message-circle-outline'
+				iconColor={color.red}
+				disabled
+			/>
+		</View>
+	)
+}
+
+export const OneToOneSettings: React.FC<ScreenProps.Chat.OneToOneSettings> = ({
+	route: { params },
+}) => {
+	const { goBack, navigate } = useNavigation()
+	const [{ flex, background }] = useStyles()
+	const { convId } = params
+	const conv = useConversation(convId)
+	const contact = useContact(conv.contactPublicKey)
+	if (!(conv && conv.type === messengerpb.Conversation.Type.ContactType && contact)) {
 		goBack()
-		return (
-			<View
-				style={{ height: '100%', width: '100%', alignItems: 'center', justifyContent: 'center' }}
-			>
-				<ActivityIndicator />
-			</View>
-		)
+		return null
 	}
 	return (
 		<ScrollView style={[flex.tiny, background.white]} bounces={false}>
-			<HeaderSettings actionIcon='upload' undo={goBack}>
-				<OneToOneSettingsHeader contact={contact} />
+			<HeaderSettings
+				action={() => navigate.chat.contactSettings({ contactId: conv.contactPublicKey })}
+				actionIcon='more-horizontal-outline'
+				undo={goBack}
+			>
+				<View>
+					<OneToOneHeader contact={contact} />
+					<OneToOneHeaderButtons />
+				</View>
 			</HeaderSettings>
-			<OneToOneSettingsBody id={contact.publicKey} />
+			<OneToOneBody />
 		</ScrollView>
 	)
 }
